% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxplotStrategybyPerformance_m2.R
\name{BoxplotStrategybyPerformance}
\alias{BoxplotStrategybyPerformance}
\title{Plot: Boxplot PV1Math by Strategy var}
\usage{
BoxplotStrategybyPerformance(
  data,
  strategy.var,
  performance.test,
  ylab.text,
  xlab.text
)
}
\arguments{
\item{data}{data}

\item{strategy.var}{strategy.var}

\item{performance.test}{performance.test}

\item{ylab.text}{ylab.text}

\item{xlab.text}{xlab.text}
}
\description{
Plot: Boxplot PV1Math by Strategy var
}
