% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cp025q01.treated}
\alias{cp025q01.treated}
\title{Treated log file and microdata for PISA 2012, CP025, Q01  (selected countries)}
\description{
Treated log file and microdata for PISA 2012, CP025, Q01  (selected countries)
}
