% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module2_SequenceOfActions.R
\name{DataArcSinebyPerformance}
\alias{DataArcSinebyPerformance}
\title{Data: Percentage in arcsine values x PISA scores by Country}
\usage{
DataArcSinebyPerformance(data, strategy.var, performance.test, country.id)
}
\arguments{
\item{data}{A \code{matrix} or \code{data.frame} where the 'strategy.var' and performance
variables are}

\item{strategy.var}{A string with the name of the strategy variable. It is "quo()" type.}

\item{performance.test}{A string with the name of the test performance variable. It is "quo()" type.}

\item{country.id}{A string with the name of the countries variable. It is "quo()" type.}

\item{plotARcSine}{Logical. Default is "TRUE".}

\item{ylab.text}{A character string giving the text of the y-axis in the plot}

\item{xlab.text}{A character string giving the text of the x-axis in the plot}

\item{ylim.vector}{A numeric vector with the limits of the y-axis in the plot}

\item{xlim.vector}{A numeric vector with the limits of the x-axis in the plot}
}
\value{
This function returns a data frame and a plot
}
\description{
This is a function that calculates the percentage in arcsine and plots it against
the PISA scores
}
