% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreqActionsSummary_m2.R
\name{FreqActionsSummary}
\alias{FreqActionsSummary}
\title{Frequency of specifics events in a variable of Actions - Summary}
\usage{
FreqActionsSummary(data, freqact.var, var)
}
\arguments{
\item{data}{A \code{matrix} or \code{data.frame} where the 'action.var'
variable is}

\item{freqact.var}{freqact.var}

\item{var}{var}
}
\value{
This function returns a \code{data.frame} with the frequency of each
  specific events from the \code{actions.search} argument and
  "Freq.Actions.Search" summary.
}
\description{
This is a function that locates specific events (using the
\code{actions.search} argument) and create new variables associate with this
strategy.
}
