% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module0_DataPreparation.R
\name{ImportSPSS}
\alias{ImportSPSS}
\title{Read SPSS process data}
\usage{
ImportSPSS(filename)
}
\arguments{
\item{filename}{character string: the name of the file or URL to read.}
}
\value{
This function returns a data frame.
}
\description{
This is a simple function that, by default, reads an SPSS data file
and save it as a data frame.
}
