% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\docType{data}
\name{m0}
\alias{m0}
\title{Module 0: Data preparation}
\format{An object of class \code{module} (inherits from \code{list}) of length 6.}
\usage{
m0
}
\description{
Module 0: Data preparation
}
\keyword{datasets}
