% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\docType{data}
\name{m1}
\alias{m1}
\title{Module 1: Time}
\format{An object of class \code{module} (inherits from \code{list}) of length 6.}
\usage{
m1
}
\description{
Module 1: Time
}
\details{
This module contains the following functions, which should be called
  by issuing "m1$<function_name>()": NumericTimeVar, PlotTimeonTaskbyVar,
  RangeTimeonTaskbyVar, SummaryTOTbyVar, TOTVar, VarTimebyID
}
\keyword{datasets}
