% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\docType{data}
\name{m2}
\alias{m2}
\title{Module 2: Actions (cognitive related)}
\format{An object of class \code{module} (inherits from \code{list}) of length 8.}
\usage{
m2
}
\description{
Module 2: Actions (cognitive related)
}
\details{
This module contains the following functions, which should be called
  by issuing "m2$<function_name>()": BoxplotStrategybyPerformance,
  DataArcSinebyPerformance, DescriptiveStrategy, FreqActionsSummary,
  PlotStrategybyCatPerformance, ScatterPlotbyPerformance, VarActionPosition,
  VarActionSearch
}
\keyword{datasets}
