% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\docType{data}
\name{Module 2}
\alias{Module 2}
\alias{m2}
\title{Module 2}
\format{An object of class \code{module} (inherits from \code{list}) of length 8.}
\usage{
m2
}
\description{
Module 2
}
\keyword{datasets}
