% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\docType{data}
\name{Module0}
\alias{Module0}
\alias{m0}
\title{Module 0}
\format{An object of class \code{module} (inherits from \code{list}) of length 6.}
\usage{
m0
}
\description{
Module 0
}
\keyword{datasets}
