% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\docType{data}
\name{Module1}
\alias{Module1}
\alias{m1}
\title{Module 1}
\format{An object of class \code{module} (inherits from \code{list}) of length 6.}
\usage{
m1
}
\description{
Module 1
}
\keyword{datasets}
