% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTimeonTaskbyVar_m1.R
\name{PlotTimeonTaskbyVar}
\alias{PlotTimeonTaskbyVar}
\title{Check response time by var}
\usage{
PlotTimeonTaskbyVar(data, tot.var, performance.item, namexlab, nameylab)
}
\arguments{
\item{data}{A \code{matrix} or \code{data.frame}}

\item{tot.var}{a vector with the total time. It is a \code{quo()} type.}

\item{performance.item}{performance.item}

\item{namexlab}{namexlab}

\item{nameylab}{nameylab}
}
\value{
This function returns a \code{data.frame} with the number of students and number de actions (min-max)
aggregated by a specific variable.
}
\description{
This is a function that reports the number of students and number de actions (min-max)
aggregated by a specific variable.
}
