% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScatterPlotbyPerformance_m2.R
\name{ScatterPlotbyPerformance}
\alias{ScatterPlotbyPerformance}
\title{Plot: Percentage in arcsine values x PISA scores by Country}
\usage{
ScatterPlotbyPerformance(
  data,
  strategy.summary,
  performance.mean,
  country.id,
  ylab.text,
  xlab.text,
  ylim.vector,
  xlim.vector
)
}
\arguments{
\item{data}{A \code{matrix} or \code{data.frame} where the 'strategy.var' and performance
variables are}

\item{strategy.summary}{strategy.summary}

\item{performance.mean}{performance.mean}

\item{country.id}{A string with the name of the countries variable. It is "quo()" type.}

\item{ylab.text}{A character string giving the text of the y-axis in the plot}

\item{xlab.text}{A character string giving the text of the x-axis in the plot}

\item{ylim.vector}{A numeric vector with the limits of the y-axis in the plot}

\item{xlim.vector}{A numeric vector with the limits of the x-axis in the plot}
}
\value{
This function returns a data frame and a plot
}
\description{
This is a function that calculates the percentage in arcsine and plots it against
the PISA scores
}
