% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module2_SequenceOfActions.R
\name{VarActionPosition}
\alias{VarActionPosition}
\title{Identify the position of specific events in a variable of Actions}
\usage{
VarActionPosition(data, action.var, actions.search)
}
\arguments{
\item{data}{A \code{matrix} or \code{data.frame} where the 'action.var' variable is}

\item{action.var}{a vector with actions. See \code{DataActionsbyID} function.}

\item{actions.search}{A character vector with the actions to be searched.}
}
\value{
This function returns a \code{data.frame} with the frequency of each specific events
from the \code{actions.search} argument and "Freq.Actions.Search" summary.
}
\description{
This is a function that locates specific events (using the \code{actions.search}
argument) and create new variables associate with this strategy.
}
