% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarTimebyID_m1.R
\name{VarTimebyID}
\alias{VarTimebyID}
\title{Extracting the start or end time}
\usage{
VarTimebyID(data, id.var, time.var, event.var, name.var.time, new.name)
}
\arguments{
\item{data}{data}

\item{id.var}{id.var}

\item{time.var}{time.var}

\item{event.var}{event.var}

\item{name.var.time}{name.var.time}

\item{new.name}{new.name}
}
\description{
Extracting the start or end time
}
