% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConcatActions_m0.R
\name{ConcatActions}
\alias{ConcatActions}
\title{Concatenate events}
\usage{
ConcatActions(data, concat.events)
}
\arguments{
\item{data}{A \code{matrix} or \code{data.frame} where the concatenate events are}

\item{concat.events}{a vector where all the events are listed. Each element of
this vector needs to be of a \code{quo()} type.}
}
\value{
This function returns a \code{data.frame} with the concatenate events in
the 'event.type' variable.
}
\description{
This function allows you to concatenate event actions from diferent variables
in a unique vector.
}
