% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStrategy_m2.R
\name{DescriptiveStrategy}
\alias{DescriptiveStrategy}
\title{Report: Descriptive statistics by strategy}
\usage{
DescriptiveStrategy(
  data,
  strategy.var,
  performance.item,
  performance.test,
  partial.credit = FALSE,
  print = FALSE
)
}
\arguments{
\item{data}{A \code{matrix} or \code{data.frame} where the 'strategy.var' and
performance variables are}

\item{strategy.var}{A character string with the name of the strategy variable}

\item{performance.item}{A character string with the name of the item
performance variable}

\item{performance.test}{A character string with the name of the test
performance variable}

\item{partial.credit}{Logical. It can be used when the item is partial credit
score.}

\item{print}{if \code{FALSE}, instead of printing table and statistics, a
list is created with the output .}
}
\value{
This function returns a report with a descriptive analysis of the
  strategy and students performance.
}
\description{
This is a function that reports a descriptive analysis of the strategy and
students performance
}
\examples{
m2$DescriptiveStrategy(cp025q01.treated, "votat", "CP025Q01", "PV1CPRO")

}
