% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logan.R
\docType{package}
\name{LOGAN}
\alias{LOGAN}
\alias{LOGAN-package}
\title{LOGAN: Log File Analysis in International Large-scale Assessments}
\description{
Overview of the LOGAN package.
}
\details{
It enables users to handle the dataset cleaning for conducting
specific analyses with the log files from two international educational
assessments: Programme of International Student Assessment (PISA) and
Programme for International Assessment of Adult Competencies (PIAAC).
}
\section{LOGAN functions}{

The LOGAN functions: ...
}

\author{
\describe{
\item{Denise Reis Costa [aut, cre],}
\item{Rolf Vegar Olsen [ctb],}
\item{Ronny Scherer [ctb],}
\item{Bjorn Andersson [ctb],}
\item{Alexandra Niculescu [ctb],}
\item{Nani Teig [ctb],}
\item{Fredrik Jensen [ctb],}
\item{Jelena Radisic [ctb],}
\item{Tove Stjern Frones [ctb]}
}
}
