% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\docType{data}
\name{m0}
\alias{m0}
\title{Module 0: Data preparation}
\format{An object of class \code{module} (inherits from \code{list}) of length 6.}
\usage{
m0
}
\description{
Module 0: Data preparation
}
\details{
This module contains the following functions, which should be called
  by issuing "m0$<function_name>()": CleanActions, ConcatActions,
  DataActionsbyID, ImportSPSS, RangeNumberActionsbyVar, TrimVar
}
\keyword{datasets}
