% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumericTimeVar_m1.R
\name{NumericTimeVar}
\alias{NumericTimeVar}
\title{Time var as a numeric vector}
\usage{
NumericTimeVar(data, vector.time)
}
\arguments{
\item{data}{A \code{matrix} or \code{data.frame}}

\item{vector.time}{vector.time}
}
\value{
This function returns a \code{data.frame} with the number of students
  and number de actions (min-max) aggregated by a specific variable.
}
\description{
This is a function that transforms a factor var time in numeric.
}
