% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module1_TimeOnTask.R
\name{NumericTimeVar}
\alias{NumericTimeVar}
\title{Time var as a numeric vector}
\usage{
NumericTimeVar(data, vector.time)
}
\arguments{
\item{data}{A \code{matrix} or \code{data.frame}}

\item{id.var}{a vector with the individuals identification. It is a \code{quo()} type.}

\item{var.group}{a vector with the group variable. It is a \code{quo()} type.}
}
\value{
This function returns a \code{data.frame} with the number of students and number de actions (min-max)
aggregated by a specific variable.
}
\description{
This is a function that transforms a factor var time in numeric.
}
