% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotStrategybyCatPerformance_m2.R
\name{PlotStrategybyCatPerformance}
\alias{PlotStrategybyCatPerformance}
\title{Check response time by var}
\usage{
PlotStrategybyCatPerformance(data, strategy.var, categ.var, namexlab,
  nameylab)
}
\arguments{
\item{data}{A \code{matrix} or \code{data.frame}}

\item{tot.var}{a vector with the total time. It is a \code{quo()} type.}

\item{var.group}{a vector with the group variable. It is a \code{quo()} type.}
}
\value{
This function returns a \code{data.frame} with the number of students and number de actions (min-max)
aggregated by a specific variable.
}
\description{
This is a function that reports the number of students and number de actions (min-max)
aggregated by a specific variable.
}
