% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTimeonTaskbyVar_m1.R
\name{PlotTimeonTaskbyVar}
\alias{PlotTimeonTaskbyVar}
\title{Check response time by var}
\usage{
PlotTimeonTaskbyVar(data, tot.var, performance.item, xlab, ylab = "Density")
}
\arguments{
\item{data}{A \code{matrix} or \code{data.frame}}

\item{tot.var}{a vector with the total time. It is a \code{quo()} type.}

\item{performance.item}{name of the item variable}

\item{xlab}{name of the plot's x-axis}

\item{ylab}{name of the plot's y-axis. Defaults to "Density"}
}
\value{
This function returns a \code{data.frame} with the number of students
  and number de actions (min-max) aggregated by a specific variable.
}
\description{
This is a function that reports the number of students and number de actions
(min-max) aggregated by a specific variable.
}
\examples{
m1$PlotTimeonTaskbyVar(cp025q01.treated, "CP025Q01.TOT", "CP025Q01",
                       xlab = "Time on task (minutes)")

}
