% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RangeNumberActionsbyVar_m0.R
\name{RangeNumberActionsbyVar}
\alias{RangeNumberActionsbyVar}
\title{Check number of students and actions by var}
\usage{
RangeNumberActionsbyVar(data, id.var, var.group, save.table = TRUE)
}
\arguments{
\item{data}{A \code{matrix} or \code{data.frame}}

\item{id.var}{a vector with the individuals identification. It is a
\code{quo()} type.}

\item{var.group}{a vector with the group variable. It is a \code{quo()} type.}

\item{save.table}{if \code{TRUE}, will save the table generated as an object
of class \code{data.frame}. Otherwise, will print the table in pandoc
format, but the object will not be saved (even if the user assigns it to an object)}
}
\value{
By default, this function returns a \code{data.frame} with the
  number of students and number de actions (min-max) aggregated by a 
  specific variable. If \code{save.table == FALSE}, however, it will print a 
  pandoc table instead.
}
\description{
This is a function that reports the number of students and number de actions
(min-max) aggregated by a specific variable.
}
\examples{
m0$RangeNumberActionsbyVar(cp025q01.treated, NewID, CNT, save.table = FALSE)

}
