% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryTOTbyVar_m1.R
\name{SummaryTOTbyVar}
\alias{SummaryTOTbyVar}
\title{Summary of time on task by var}
\usage{
SummaryTOTbyVar(data, tot.var, performance.item)
}
\arguments{
\item{data}{A \code{matrix} or \code{data.frame}}

\item{performance.item}{a vector with the group variable. It is a \code{quo()} type.}

\item{start.time}{a vector with the individuals identification. It is a \code{quo()} type.}

\item{end.time}{a vector with the group variable. It is a \code{quo()} type.}
}
\value{
This function returns a \code{data.frame} with the number of students and number de actions (min-max)
aggregated by a specific variable.
}
\description{
This is a function that reports the number of students and a summary of time on task
aggregated by a specific variable.
}
