% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryTOTbyVar_m1.R
\name{SummaryTOTbyVar}
\alias{SummaryTOTbyVar}
\title{Summary of time on task by var}
\usage{
SummaryTOTbyVar(data, tot.var, performance.item, na.rm = FALSE)
}
\arguments{
\item{data}{A \code{matrix} or \code{data.frame}}

\item{tot.var}{a vector with the time on task.}

\item{performance.item}{a vector with the group variable. It is a
\code{quo()} type.}

\item{na.rm}{remove missing data in `performance.item`? Default is `FALSE`}
}
\value{
This function returns a \code{data.frame} with the number of students
  and number de actions (min-max) aggregated by a specific variable.
}
\description{
This is a function that reports the number of students and a summary of time
on task aggregated by a specific variable.
}
\examples{
# Data preparation
df <- cp025q01
df$NewID <- paste0(df$cnt, "-", df$schoolid, "-", df$StIDStd)
trim.vars <- c("event", "event_type", "diag_state")
df.trimmed <- m0$TrimVar(df, trim.vars)
library(rlang)
concat.events <- c(quo(event), quo(event_type))
df.conc <- m0$ConcatActions(df.trimmed, concat.events)
clear.events <- c("ACER_EVENT_" = "", "_NULL" = "")
df.clean <- m0$CleanActions(df.conc, event.type, clear.events)
time.vars <- c("cnt", "schoolid", "StIDStd", "NewID")
df.start <- m1$VarTimebyID(df.clean, NewID, time, new.event.type,
                           "START_ITEM",
                           "CP025Q01.START")[c(time.vars, "CP025Q01.START")]
df.end   <- m1$VarTimebyID(df.clean, NewID, time, new.event.type,
                           "END_ITEM",
                           "CP025Q01.END")[c(time.vars, "CP025Q01.END")]
df.dataAct <- m0$DataActionsbyID(df.clean, NewID, new.event.type,
                                 "CP025Q01.ACTIONS")[c(time.vars, "CP025Q01.ACTIONS")]
df.time <- dplyr::left_join(df.start, df.end, by = time.vars)
df.timeActions <- dplyr::left_join(df.time, df.dataAct, by = time.vars)
df.dataAct <- m1$TOTVar(df.timeActions, "CP025Q01.START", "CP025Q01.END",
                        divBy = 60, tot.var = "CP025Q01.TOT")
micro <- pisa
names(df.dataAct)[1:2] <- c("CNT", "SCHOOLID")
df.complete <- dplyr::left_join(df.dataAct, micro,
                                by = c("CNT", "SCHOOLID", "StIDStd"))
# Function demonstration
m1$SummaryTOTbyVar(df.complete, "CP025Q01.TOT", "CP025Q01", TRUE)
}
