% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryTOTbyVar_m1.R
\name{SummaryTOTbyVar}
\alias{SummaryTOTbyVar}
\title{Summary of time on task by var}
\usage{
SummaryTOTbyVar(data, tot.var, performance.item, na.rm = FALSE,
  save.table = TRUE)
}
\arguments{
\item{data}{A \code{matrix} or \code{data.frame}}

\item{tot.var}{a vector with the time on task.}

\item{performance.item}{a vector with the group variable. It is a
\code{quo()} type.}

\item{na.rm}{remove missing data in `performance.item`? Default is `FALSE`}

\item{save.table}{if \code{TRUE}, will save the table generated as an object
of class \code{data.frame}. Otherwise, will print the table in pandoc
format, but the object will not be saved (even if the user assigns it to 
an object)}
}
\value{
This function returns a \code{data.frame} with the number of students
  and number de actions (min-max) aggregated by a specific variable.
}
\description{
This is a function that reports the number of students and a summary of time
on task aggregated by a specific variable.
}
\examples{
m1$SummaryTOTbyVar(cp025q01.treated, "CP025Q01.TOT", "CP025Q01")

}
