% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrimVar_m0.R
\name{TrimVar}
\alias{TrimVar}
\title{Trim variables}
\usage{
TrimVar(data, trim.vector)
}
\arguments{
\item{data}{dataset}

\item{trim.vector}{vector of variables on the dataset to be trimmed}
}
\value{
This function returns a vector removing trailing and leading spaces
inside the original vector.
}
\description{
\code{TrimVar()} is a function that allows you to remove whitespace inside the
strings of a vector.
}
\examples{
head(m0$TrimVar(cp025q01, "event"))

}
