% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarActionSearch_m2.R
\name{VarActionSearch}
\alias{VarActionSearch}
\title{Frequency of specifics events in a variable of Actions}
\usage{
VarActionSearch(data, action.var, actions.search)
}
\arguments{
\item{data}{A \code{matrix} or \code{data.frame} where the 'action.var'
variable is}

\item{action.var}{a vector with actions. See \code{DataActionsbyID} function.}

\item{actions.search}{A character vector with the actions to be searched.}
}
\value{
This function returns a \code{data.frame} with the frequency of each
  specific events from the \code{actions.search} argument and
  "Freq.Actions.Search" summary.
}
\description{
This is a function that locates specific events (using the
\code{actions.search} argument) and create new variables associate with this
strategy.
}
\examples{
\dontrun{
  df.dataAct <- m2$VarActionSearch(df.timeActions,
                                   "CP025Q01.ACTIONS", #'"1_apply")
  table(df.dataAct$freq.1_apply)
}
}
