% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarTimebyID_m1.R
\name{VarTimebyID}
\alias{VarTimebyID}
\title{Extracting the start or end time}
\usage{
VarTimebyID(data, id.var, time.var, event.var, name.var.time, new.name)
}
\arguments{
\item{data}{data frame}

\item{id.var}{vector of unique identification}

\item{time.var}{vector with the time variable}

\item{event.var}{vector with the events}

\item{name.var.time}{name of the time string to filter (ex.: "START_ITEM" or
"END_ITEM")}

\item{new.name}{name of the output variable}
}
\value{
a data frame with `time` replaced with `new.name`. The variable
  `event.var` is dropped.
}
\description{
Extracting the start or end time
}
\examples{
\dontrun{
  m1$VarTimebyID(df.clean, NewID, time, new.event.type, "START_ITEM",
                 "CP025Q01.START")
}
}
